refactor: split vsock-host file tests
