+++ crates/vsock-host/src/tests/file/mod.rs
@@ -1,0 +1,5 @@
+mod copy;
+mod read;
+mod support;
+mod write;
+mod write_chunked;
